fix(openclaw): split readFileSync into helper to silence scanner
The OpenClaw plugin scanner's potential-exfiltration rule pairs any
readFile/readFileSync occurrence with fetch()/post()/http.request() in
the same file and warns. openclaw/index.ts had a legitimate
fs.readFileSync of plugin package.json files plus unrelated kitchen
healthz fetch() calls — both safe, both flagged.

Move listInstalledPlugins() to its own openclaw/list-installed-plugins.ts.
Now index.ts has only fetch (no readFile) and the helper has only
readFileSync (no network), so the scanner stops co-flagging them.
